app/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
gradio==5.35.0
httpx==0.27.2
loguru==0.7.3
pillow==11.0.0
numpy==1.26.4
prometheus-client==0.21.1
pydantic-settings==2.6.1
tritonclient[http]==2.48.0
opentelemetry-sdk==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
